Before I go any further with the GUI . . .
I need to work out a method of storing settings . . . for example, the
location of the aesalon gdb interface executable, or the xterm to use . . .
you get the picture.

I have no clue on how to do this . . . perhaps a map of strings? And load
configuration information from ~/.aesalon/qt_gui.cfg or something . . .