Fixed a bug in MallocObserver::notify().
It was a bit of a re-entrant problem, combined with a gdb error . . .
Basically, when the gdb interpreter gets '-exec-finish', it finishes the
execution of the current frame. However, if there is a breakpoint on the same
place that the '-exec-finish' goes to, then the message sent out is not a
"function-finished", it is a "breakpoint-hit". That, and the observers were
added in an incorrect order . . .

Now, I have to do something similar for free(). Unfortunately, the difference
here is that I don't have a nice return value to watch for . . . so I don't
know how to watch for it . . .