A very unwieldy solution to the speed problem . . .
. . . that might not even be a solution. According to an excellent blog post
(http://t-a-w.blogspot.com/2007/03/how-to-code-debuggers.html), ptrace can be
used to get the content of a program's memory at a specific address. This
means that, given enough time, I can write my own data-gatherer, by parsing
ELF and DWARF information from an executable. In essence, I'll end up
duplicating parts of GDB, but only the parts I want. It'll be tricky, no doubt
about it, but it might work . . .

We'll see. I'll keep looking for a better solution.