Very well. ptrace-based information gathering it is.
This means that aesalon will be tied to UNIX-based systems . . . oh well. I
can think of worse problems that could occur. Such as the opposite . . .

I'm going to create an overload library again, I think, for the allocation
information. In other words, here's where the data is coming from:

- Allocation information from the overload library, sent along a UNIX pipe
- Registers and raw memory values from ptrace
- Variable names and stack positions from DWARF parser

Which brings to mind the fact that the Platform namespace has precious little
platform-specific code . . . this overload/ptrace/DWARF combination is going
to be very platform-specific, unlike the current design that uses gdb. Which
means that I'm going to have to worry about 32-bit and 64-bit disreprancies.

This commit is basically a backup of the old gdb-based system in case I need
it later . . .