Added in a few small functions to random classes.
But that's not the important part at the moment. The important part is how I
am going to implement the DWARF parser. According to the DWARF 3
specification, DWARF 3 "is structurally upward compatible with DWARF version 2
execpt as follows [. . .]". The only interesting point in the
incompatibilities list is this one:
- References that use the attribute form DW_FORM_ref_addr are specified to be
four bytes in the DWARF 32-bit format and eight bytes in the DWARF 64-bit
format, while DWARF Version 2 specifies that such references have the same
size as an address on the target system (see Sections 7.4 and 7.5.4).

Since this is only a minor change, implementing a DWARF 3 parser should work
for aesalon, as it appears that g++ generates debugging info in DWARF 2.

So . . . next up, the beginnings of the DWARF storage tree.