The ELF parser is almost working.
The only issue is figuring out which section is .shstrtab . . . which, of
course, I can't figure out without knowing the names of the sections . . .
which you can't know without knowing which is .shstrtab. Ergo, circular logic.

Wonder how I can solve that. It appears to always be the third-last section,
but I don't want to rely on that . . .