I've been reading the DWARF documentation . . .
And I've come to a conclusion: the DWARF parser is going to be rather large.
I need to completely read the entire documentation before I go very much
farther, and the DWARF 3 standard specification is 267 pages. Not exactly your
late-night light reading.

However, in the meantime. The GUI crashes for some reason when you try to
quit. I should probably fix that.