Sort-of implemented history for the New Program dialog.
However, a "Session"-based system seems more prudent, to be honest. Basically,
this means that you create a new "Session", which has a given executable path.
You then store the history of different arguments etc. that you pass to the
executable. It seems like a better idea than the mix-and-match style that is
the current system . . .