Scratch the idea about the minimalistic disassembler.
It'll be more trouble than it's worth to try to do that; a full disassembler
will probably, strangely enough, be easier. I'm also going to try to implement
everything hard-coded to the platform, rather than use an XML file. I do
realize it's probably an idea I'll regret, but oh well. That's how these
things work.