So, there are two ways I can do the disassembly . . .
I can have a gigantic set of if() statements that are hardcoded, e.g. opcode
0x00 is "MOV <8-bit register>, <constant>" . . .

Or I can put it into a data file and parse it that way, e.g. softcode it. I
can take advantage of the ability to pack on files into a single ELF file, so
that no external data files are actually required . . . but that would be
annoying, as previously mentioned.

Oh well, I can always change it later.