I wonder if I actually need a full disassembler, after all . . .
Think of it this way -- very little actually needs to be understood about the
instruction being executed besides what memory it changes and where it gets
that value from . . . if it is, for example, a mov instruction, then it copies
the "taint" to the new value. If it is a add instruction, then it checks to
ensure that the new value is still within the same block of dynamically-
allocated memory.

That is a gross oversimplification, but I think it might actually work . . .

We'll see.