A thought on the representation of assembly . . .
I don't actually need to know what each instruction is -- all I need to know
is what memory or registers it changes. If I represent a "mov" instruction as
one that changes memory, source is the first operand, changed memory is the
second operand, it would make things much easier.