Began an observer-based system for handling breakpoints.
Basically, whenever a breakpoint is hit, a reverse iteration starts through
the list of breakpoint observers. A reverse iteration, since the last item in
the vector is typically the first one to be looking for. I'm not entirely
certain how long it will take to iterate through a vector with >50000 items,
though . . .

Whatever. I will quote PoppaVic again: "Make it work, then make it work right,
then make it fast or small (choose one)."