Added SegfaultObserver to the PTrace:: namespace.
Basically, it looks for segfaults (and eventually bus errors and other signals
that cause the abortion of a program), prints a warning if it finds one, then
tells the ProgramManager that the program is no longer running.

Still no clue on why the breakpoint on malloc() is causing a segfault. Reminds
me, I need to move the code that actually places the breakpoint elsewhere.