Well, fixed one bug. At least the child isn't segfaulting anymore . . .
TrapObserver is setting the breakpoint on main(), just to make things easier
for the moment. However, Portal::read_memory() is claiming that there is "No
such process". From the man page for ptrace, ESRCH is set when "The specified
process does not exist, or is not currently being traced by the caller, or is
not stopped (for requests that require that)."

Since it should be stopped (since a SIGTRAP was just caught), and the process
exists . . .

Just my programming, as usual, I suppose.