A few cosmetic changes to the GUI, in an effort to improve it slightly.
Now, the real fun begins in the GUI. This is what I am currently thinking for
the GUI: Network threads (one per socket) listen to sockets and push events
onto a queue. Every second or so, (user-definable) each ActiveSession instance
updates itself from the queue, and therefore updates the views available.

If it sounds needlessly complex, that's because it is.