Removed a huge amount of code from platform that is no longer being used.
This is what I am thinking for communication between the monitor and the
GUI(s): the monitor traps a malloc call, say with a size of 32. MallocObserver
generates an event, a new block event, with an address of whatever it was, and
the given size. The event is then pushed onto the EventQueue, ready for
transmitting. A local copy is also kept, as a representation of the memory --
useful for new clients -- and perhaps logged (the time is also added when the
event is logged, of course).

The transmission is done via a TCP network socket -- UDP is too unreliable,
and I don't want to use RUDP due to its relative unportability -- to each and
every listening GUI. This will be a lot of information, unfortunately, so it
will have to be compressed very well. The listening TCP socket (which will be
a GUI instance) will then proceed to deserialize the event, push it onto its
own local queue, and then process the even as so.

A fair design, I think.