Well, as I knew it would, $rbp-40 is coming back to haunt me.
It would appear that the return address of malloc changes depending on where
and how it is called. $rbp-40, for malloc_test, just happened to be the right
address. Grr.

libc is defying all I know of assembly, and I don't think that's a good thing.
I suppose it is because libc is probably the most-heavily optimized piece of
software in the world . . .

It's research time.