Implemented support for actually executing programs in the GUI.
Now, assuming the configuration file has the correct aesalon-path set (which
there is currently no way to set graphically, you have to edit the
configuration file manually . . .), and you tell the GUI to execute a non-
interactive program (since, of course, the console output is invisible; it's
being redirected to nowhere), it will do so, and display the statistics from
the program in question -- number of allocations, deallocations, and active
(leaked) blocks.

Also added support for unknown signals in the monitor. Now, instead of
crashing when a non-SIGSEGV or SIGTRAP signal is caught, it instead passes it
onto the program being monitored.