A few small changes in the GUI, such as supporting abritrary port numbers.
Now, for the taint-tracking, otherwise known as the hard part.

Here's an example snippet of assembly (assume that malloc() was just called):
    mov     rax, qword [rbp-16]
    xor     rax, rax

Upon returning from malloc (MallocObserver will have to set a TaintTracker
observer on the breakpoint it sets . . .), the TaintTracker will mark the
register RAX as tainted. On the mov line above, it will notice that the value
of a tainted register is being put into the address [rbp-16] (presumably a C-
style pointer, in this case), and mark that memory address tainted. Then, the
next line zeroes out rax, therefore removing the taint.

One issue that occurs to me now is something like this:
    xor     rax, 0xff
    xor     rax, 0xff

In this case, after the second xor, rax is re-tainted, but the taint tracker
would not notice that . . . Though I very much doubt that would ever occur,
excepting trying to fool aesalon, that is . . .

The main question here is something along these lines: where to place
breakpoints? Or should aesalon single-step through the entire program? Because
in the first example given, I, as a human, can tell you what is happening, but
place another fifteen unrelated instructions between those two, and I'd get a
little (more than a little, actually) lost. Single-stepping would solve that,
but it also raises more problems, mainly speed. That, and a program that uses
no dynamically-allocated memory should not be affected by aesalon.