Here's the process that I'm going to use for the taint tracking . . .
At least, I think I'm going to use it. In essence, it's a bit of a hackish
approach to it, but it should get around the initialization time issue at
least. First off, MallocObserver adds a TaintObserver to the return-address
breakpoint it places. When the TaintObserver's handle_breakpoint is called,
then it marks the register rax as tainted. When it does so, it places more
TaintObserver'd breakpoints wherever the value of rax is copied into memory or
other registers. In turn, those TaintObservers will place more breakpoints
where those addresses are used . . . and so on. This gets around another
issue, regarding register offsets, where the value of the register will
change, so [rbp-16] is not always pointing to the same place . . .

Sounds good, at least in principle . . .