A few things that need thinking about . . .
First of all, I wonder if I can embed a Konsole inside each ActiveSession, as
that would allow multiple programs to be run from within the GUI. That might
be a KDE-only feature (not a Qt feature), but it should be possible, I would
think. I'll take a look at the Qt documentation, I suppose.

Next off: how is the data gathered going to be represented? The "flat" views,
while easy, are not very useful . . . a 3D view of the memory currently being
used could be nice -- X-axis is allocated blocks, Y-axis is the total
allocated memory, Z-axis is time, in seconds . . . or, a tree view, with
zooming etc. (Think UML diagrams . . .)

Also, aesalon currently isn't working properly when run on itself -- it is not
reporting all of the allocations or frees or something similar; it claims
there are 32 allocations, 33 frees. I'm going to hazard a guess it's a result
of aesalon being written in C++, but I'm not sure. I'll try the monitor on
other C++ programs, I suppose.