Properly implemented socket-waiting.
Now, when a program is executed, aesalon properly establishes a connection
even when the monitor doesn't listen on a port right away. This means one can
set up a GUI to try connecting to a port that doesn't have a monitor yet, but
will soon, as to catch all of the events waiting.