Fixed the Configuration's file-selection system.
Added --wait-for <n> in the monitor, which, when combined with --wait, waits
for a specific number of TCP connections before it begins executing.

Also added a configuration entry, xterm-path, or the path to the xterm to
launch the monitored program inside. This is until I can figure out how to
embed the console session inside the GUI; in fact, I may keep it this way.