Well, I've come to a conclusion . . .
And that conclusion is: platform needs to go. It is no longer required . . .
or, at least, most of it is no longer required. Taking a look at the source
files in src/platform/, the only source files being used are: Event and
EventQueue. Well, and ArgumentList. But ArgumentList can be put into misc, and
so can Event and EventQueue. In essence, I can redefine the roles of each
namespace: Monitor is all of the nitty-gritty stuff, the actual data
collection, GUI is the fancy displaying-of-data, and Misc is anything that is
common between the previous two. Platform is, so to speak, out.

On a side note, with this commit, aesalon passes 6000 lines.