Began removal of platform/. The Monitor:: namespace is completely converted.
The GUI, I suppose, is next. However, there are a few things that occured to
me as I was doing this part of the conversion: the Misc:: namespace is
actually not required anywhere in the GUI -- I checked -- and, as such, the
GUI could very well use its own buildsystem . . . what I'm thinking is
splitting the src/ directory into monitor/ and gui/ (both in the project
root), and building the executables that way . . . makes sense, I suppose.