Continued ripping up the source.
Aesalon's source code is now split into two sections, monitor/ and gui/.
Something I'll have to be careful about, though, is the communication code --
serialization and deserialization have to be identical, otherwise the whole
shebang will collapse . . .

monitor/ will be re-organized/re-designed/re-written/re-your-choice-of-word
very differently than it used to be. Numero uno: away with SmartPointers! If I
simply do a little bit of thinking, I can eliminate all of the pesky slowdowns
and problems that SmartPointers introduce. (If a function is passed a pointer,
then it is the responsibility of the parent to free the pointer, etc.)

Numero duo: the source code will be refactored, and much of what is currently
in misc/ will be changed. All of the Singletons (besides Initializer) will be
removed (pointers to everything will be accessible via Initializer). The Event
classes will be moved into their own namespace. Likewise with the Exception
classes. There will remain, of course, a "misc" subdirectory for classes such
as String and StreamAsString, which don't particularily belong in any one
given place.

gui/ will be likewise refactored, and sections of code will be re-written to
make them more Qt-like. I want the GUI to be as portable as possible, since
there is very little reason not to . . . if the monitor isn't going to be
portable, then I can at least make the GUI portable, I suppose.

That should be more than enough to chew on for the moment.