Pretty much finished refactoring the monitor.
The only major thing remaining in the monitor is to re-implement the argument
parser. Beyond that, I don't think anything else needs to be done, besides
beginning to implement the reference tracking . . .

It truly is amazing how much extraneous source aesalon had. The monitor has
gone from 4759 lines to a mere 3281 lines, or 1478 lines smaller now . . .