Replaced the GUI's storage system with a better one.
The new system is much more complex, but it will enable histories (e.g.
restoring the status of the program to a specific point in time), and also
removes a large amount of very, very ugly code from a number of places -- in
particular, removes the requirement of a lot of dynamic_cast<>s by using a
command design system.