Changed a few methods in the monitor to speed things up a little.
While it is faster than it was before, something drastic still needs to be
done. Perhaps creating a shared memory map instead of using
Portal::read_memory() everywhere?

Note that ServerSocket::accept_connections() seems to be malfunctioning
slightly -- the select() call is claiming there is a connection waiting, when
there is in fact not. For the moment, the call to accept_connections() is
simply commented out.