I think that shared memory maps (shms) are the way to go, instead of ptrace().
Well, of course ptrace() is required for some things, but it will be faster to
use a memory map instead of ptrace for reading and writing the memory of the
child.

Now, how to go about this . . . I can set a breakpoint on the main() function,
as I currently do, and replace the first few bytes with a jmp instruction, one
that jumps to a location in memory that is currently unused (e.g. the
beginning of the heap), whereupon that code will create the memory maps . . .

One difficulty is actually loading the memory maps. The shm functions are
contained within the librt shared library . . . if LD_PRELOAD is set to load
/lib/librt.so.1, then it will be available, and I can use my libc parser to
get the offset of whatever function I need. Works, I think . . . well, let's
see.

Here's what the injected code needs to do:
+ Call shm_open() with a few pre-determined flags and a C-style string that
    will be loaded into memory via the monitor (beginning of the heap, can use
    rip as an offset to find the data required)
+ Call ftruncate() with the memory size wanted. This might be the tricky part,
    as I don't know how well it would respond to my trying to create a memory
    map of size 0x7fff00000000 or so . . . I'll have to try that, I guess.
+ Call mmap() to create the actual map itself.

Then, the monitor needs to set the permissions on the memory map so that no
one else can read the shm. Once the execution has finished, the monitor needs
to unlink the shm, etc. as well . . .