I've been thinking about the way the GUI stores information . . .
. . . and come to a conclusion: the current setup is too inefficient. Yes, it
does an admirable job of being extensible to the Local Group, and also of
providing accurate recall to any millisecond in time of the program's
execution, but is that really required? Already, even for something simple,
the system is far too slow . . .

So, an alternative design (still using a multiple-feedback queue-style system)
could be something like this . . .

When an event is processed, it is immediately applied to the "current"
snapshot, e.g. the current status is always up-to-date. That part is the same
as the current setup, but here's where the interesting part is: as a block is
allocated, it is given 25 bytes on a huge mmap()ed (or something similar)
region of memory, from the end of the currently-reserved space. Then, when a
snapshot is pushed, it stores the *differences* since the last snapshot (or,
in a few cases, a full update, for speed reasons . . .), but instead of
storing pointers, it stores an offset (which will be, say, four bytes) that
can be used to retrieve the block . . . or, a snapshot . . .

Of course, each snapshot has a second list of blocks that have been removed,
but that's a different story. The fun bit about this is that when references
are implemented, it is simply the work of adding another two lists, or even
re-using the same lists that are used for blocks . . .

This means that there will be two "current" snapshots, the first is simply the
full list of all the blocks/references/whatever other data, and the second is
the differences since the last snapshot, periodically emptied.

Fun, isn't it?