Implemented snapshot-saving in the GUI.
The snapshots (rather, their offsets) are saved on a QList, ready to be
accessed whenever required. Also added a nice GUI configuration item in
SessionEditor that allows a configurable snapshot interval, in milliseconds.

Now, back to the monitor . . . I'll see what I can do to speed it up. Perhaps
using LD_PRELOAD instead of ptrace() to overload the functions would be a good
idea? It would certainly speed it up . . .