Attempted to re-implement the BlockView . . .
But something very weird is going on with the ActiveSessionMemoryStorage
class . . . something that I can't pin down. It's probably just a small
mistake somewhere, but I can't track it down at the moment, as the text is
starting to blur in front of my eyes at the moment.

I'll try again tomorrow when I can actually see properly.