Added simple calloc test to the tests/ subdirectory.
So, what am I going to work on next? The monitor and the GUI both need lots of
attention:
+ The monitor needs to be sped up by a factor of about 20-50, some of which
    can be accomplished via an overload library and code injection . . .
+ The GUI needs an ActiveSession redesign -- the current system is hacked
    together too badly to be very extensible in the future . . .

Well, I suppose I should work on the monitor, since the GUI is rather useless
without the monitor working properly. So, monitor it is . . .

Now, what to do in the monitor? I suppose I'll start off with replacing the
allocation-information gathering with an overload library again.