Successfully replaced the ptrace-based allocation gathering with LD_PRELOAD.
The speed results are much more promising now, a mere four seconds for
fib_benchmark -- rather than ten seconds -- but are still not (in my opinion)
enough . . . it's still a good 250 times slower. But it's a definite
improvement, that much is certain.

Note that you currently have to pass the full path of the overload library to
the monitor, and no symlinks allowed. Which gets rather annoying, so I think
I'll see what I can do about fixing that . . .