An eureka moment that should never have had to happen . . .
As far as the data views go, rather than implementing a whole group of
visualizations, I think I'll just go with two major ones: a table and a graph.

The graph can be configured (two, three axis), with a number of data sources,
the control of which is always time. The y-axis can be configured as required.
The table is simply a table which can acquire data from a number of sources,
such as the blocks, the references, etc.

Any number of table/graph tabs can be spawned off, to provide a nice way of
displaying the information across multiple screens. The primary difference
(besides the obvious display differences . . .) is how each receives
information: graphs act upon a range of pre-calculated data that is abstracted
somewhat from the actual data (or, in a less confusing manner, calculated
statistics from given data), whereas tables etc. act upon the actual gathered
data itself.

However, that's for the future . . . the tree needs to be implemented first.