Successfully implemented AllocEvent and FreeEvent in the GUI.
A binary-search-style tree is now created for blocks, and said blocks added to
the correct leaf, with a configurable tree depth (default is 52, 64 - 12,
since linux page sizes are 4K . . . doesn't really make much sense, but it's
about the right size . . .). In theory (I haven't tried it yet), it should
also properly create the history that is required for aesalon's GUI to perform
all of its work.

It's taken long enough to do this . . . now, time for . . . something else.

Not sure what . . . graph/table framework in the GUI, I suppose.