Here's how the graph system is going to work . . .
There are a few Graph instances (each one is a different QWidget, they can
display the same data if they wish to), which all reference a
GraphStorageEngine (or something similar) that stores already-calculated
statistics. Else, the data is requested on a per-snapshot basis, and the data
is calculated, then stored inside the StorageEngine. A StorageEngine is
specific to the current Session.

GraphData uses a sort of command-based design pattern: it has a member called
render_coord(), which returns a qreal representing the "height" of that
particular element. The GraphRenderer then creates a graph from that
information. The QImage is generated inside a separate rendering thread, then
displayed once it is fully complete (or it could be split into sections with
multiple threads generating the graph . . .). Something like that, anyhow.

Good enough for now.