The data requesting, etc. is a little more complex than I thought it was . . .
The data needs to be sent along to not only the GraphDataEngine instance, but
also along to whatever rendering thread requested the data in the first place.

This means that I need to pass the DataRequest along . . . so, in other words,
the rendering thread needs to push 200-odd requests onto the GraphDataEngine's
queue, which will then turn around and either give them right back to the
render thread if the data is cached, or pass them along to the DataThread,
where they will be processed, sent back to the GraphDataEngine, and then sent
to the render thread. Sounds a bit like the description of a Vogon . . .

Very well . . . I just can't help but think that there must be a better way
to do this . . .