Just wondering where the GraphDataEngines should be stored . . .
They don't belong in the StorageThread, but I think that's where they're going
to go. They do not belong inside the RenderThreads, that is, at least, if
there are going to be multiple RenderThreads, which I suspect is the case.

I don't want to introduce another thread into the mix, thank you very much.

Also, I have an idea for a nice direct data-representation view: basically, a
view of the program's blocks: the vertical axis is time, the horizontal axis
is address. A block of memory is drawn as a rectangle, with its horizontal
size being the size of the block, and its vertical size how long the block was
allocated for. Large gaps where there are no allocated blocks, will, of course
be compressed. Sounds more useful than a statistical view, doesn't it?